+++ src/views/AuditDetail/AuditDetail.vue
@@ -77,6 +77,40 @@
 					<hr role="presentation" class="separator" />
 
 					<section class="section-logs">
+
+						<h2>{{ $t('auditDetail.logs') }} ({{auditLogTotal}})</h2>
+
+						<div class="form-row">
+							<div class="form-column">
+								<div class="form-block">
+									<div class="select">
+										<select id="level-select"  name="level-select" @change="levelChange($event)" required>
+											<option value="" disabled>{{$t('auditDetail.selectLogLevel')}}</option>
+											<option v-for="level of levels" :key="level" :value="level">{{level}}</option>
+										</select>
+									</div>
+								</div>
+							</div>
+							<div class="form-column">
+								<div class="form-block">
+									<button
+										type="button"
+										:class="firstToLast ? 'btn btn--default-inverse btn--icon' : 'btn btn--default btn--icon'"
+										@click="reverseLogsOrder()"
+										aria-pressed="true">
+										{{ $t('action.sortLogs') }}
+										<icon-base-decorative v-if="firstToLast">
+											<icon-close/>
+										</icon-base-decorative>
+
+										<icon-base-decorative v-else>
+											<icon-checked/>
+										</icon-base-decorative>
+									</button>
+								</div>
+							</div>
+						</div>
+
 						<logs-list
                         :audit="audit"
                         :auditLogs="auditLogs"
@@ -85,12 +119,12 @@
                         :element-by-page="auditLogPageSize"
                         :total-elements="auditLogTotal"
                     	/>
-
 						<pagination
 							:current-page="auditLogCurrentPage"
 							:total-pages="auditLogTotalPage"
-							@changePage="(page) => {loadAuditLogs(page, auditLogPageSize)}"
+							@changePage="(page) => {loadAuditLogs(page, auditLogPageSize, !firstToLast, this.levelsToDisplay)}"
 						/>
+						
 					</section>
                 </div>
 			</Tab>
@@ -104,14 +138,15 @@
 <script>
 import Breadcrumbs from "../../components/Breadcrumbs";
 import Tab from '../../components/Tab';
-import Tabs from '../../components/Tabs';
+import Tabs from '../../components/Tabs';search bar for users/contrats adapted for pagination
 import Logs from './LogsList';
 import Synthesis from './Synthesis';
 import LogsList from "./LogsList";
 import PageList from "./PageList";
 import Pagination from "../../components/Pagination";
 import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
-
+import IconChecked from '../../components/icons/IconChecked'
+import IconClose from '../../components/icons/IconClose'
 	export default {
 		name: "auditDetail",
 		components: {
@@ -123,7 +158,9 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
 			Tabs,
 			Logs,
 			Synthesis,
-			IconBaseDecorative
+			IconBaseDecorative,
+			IconChecked,
+			IconClose
 		},
 		metaInfo() {
 			return {
@@ -146,6 +183,9 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
           		project: null,
 				pages: [],
 				auditLogs: [],
+				firstToLast: false,
+				levels: [],
+				levelsToDisplay: ['INFO','WARNING','ERROR'],
 				parameters: {
 					mainReference: null,
 					browser: null,
@@ -175,12 +215,11 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
 		created() {
 			this.sharecode = this.$route.params.sharecode;
 			this.getProject();
+			this.getLogLevels();
 			this.refreshPages();
 			this.timer = setInterval(this.refreshPages, 3000);
-
-            this.loadPages(this.pageCurrentPage, this.auditPagePageSize, this.search);
-			this.loadAuditLogs(this.auditLogCurrentPage, this.auditLogPageSize);
-
+      this.loadPages(this.pageCurrentPage, this.auditPagePageSize, this.search);
+			this.loadAuditLogs(this.auditLogCurrentPage, this.auditLogPageSize, !this.firstToLast, this.levelsToDisplay);
 			this.getParameters();
 		},
 		beforeDestroy () {
@@ -264,9 +303,8 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
 						console.error(error);
 					}
 				);
-
-                this.loadPages(this.pageCurrentPage, this.auditPagePageSize, this.search);
-                this.loadAuditLogs(this.auditLogCurrentPage, this.auditLogPageSize);
+        this.loadPages(this.pageCurrentPage, this.auditPagePageSize, this.search);
+        this.loadAuditLogs(this.auditLogCurrentPage, this.auditLogPageSize, !this.firstToLast, this.levelsToDisplay);
 			},
 
             loadPages(page, size, name){
@@ -285,12 +323,14 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
                 )
             },
 
-            loadAuditLogs(page, size){
-                this.auditLogService.findByAuditId(
+            loadAuditLogs(page, size, asc, levelsToDisplay){
+                this.auditLogService.findByAuditIdFiltered(
                     this.$route.params.id,
                     this.sharecode,
                     page,
-                    size,
+					size,
+					asc,
+					levelsToDisplay,
                     (auditLogsPage) => {
                         this.auditLogCurrentPage = page;
                         this.auditLogs = auditLogsPage.content;
@@ -310,7 +350,7 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
 					this.getAudit();
 				}
 			},
-
+      
 			fireAriaLive(){
 				clearTimeout(this.timer)
 				this.timer = setTimeout(this.populateAriaLive, 1000)
@@ -319,7 +359,35 @@ import IconBaseDecorative from '../../components/icons/IconBaseDecorative';
 			populateAriaLive(){
 				this.loadPages(0, this.auditPagePageSize, this.search)
 				this.liveMsg = this.pages.length + ' ' + this.$i18n.t('auditDetail.pagesNb')
+      },
+      
+			reverseLogsOrder(){
+				if(this.firstToLast == true) {
+					this.firstToLast = false
+				} else { this.firstToLast = true }
+				this.loadAuditLogs(this.auditLogCurrentPage, this.auditLogPageSize, !this.firstToLast, this.levelsToDisplay);
+			},
+
+			getLogLevels(){
+				this.auditLogService.getLogLevels(
+					levels => {
+						this.levels = levels
+					},
+					err => console.error(err)
+				)
+			},
+
+			levelChange(event) {
+				if(event.target.value === 'INFO'){
+					this.levelsToDisplay = ['INFO', 'WARNING', 'ERROR']
+				}else if(event.target.value === 'WARNING'){
+					this.levelsToDisplay = ['WARNING', 'ERROR']
+				}else if(event.target.value === 'ERROR'){
+					this.levelsToDisplay = ['ERROR']
+				}
+				this.loadAuditLogs(this.auditLogCurrentPage, this.auditLogPageSize, !this.firstToLast, this.levelsToDisplay);
 			}
+
 		},
 		computed: {
 			breadcrumbProps(){